+++ docs/backend-system/index.md
@@ -10,4 +10,4 @@ description: The Backend System
 
 The new backend system is released and ready for production use, and many plugins and modules have already been migrated. We recommend all plugins and deployments to migrate to the new system.
 
-You can find an example backend setup in [the `backend-next` package](https://github.com/backstage/backstage/tree/master/packages/backend-next).
+You can find an example backend setup in [the `backend` package](https://github.com/backstage/backstage/tree/master/packages/backend).
