+++ package.json
@@ -24,7 +24,7 @@
     "clean": "backstage-cli repo clean",
     "create-plugin": "echo \"use 'yarn new' instead\"",
     "dev": "yarn workspaces foreach -A --include example-backend --include example-app --parallel -v -i run start",
-    "dev:next": "yarn workspaces foreach -A --include example-backend-next --include example-app-next --parallel -v -i run start",
+    "dev:next": "yarn workspaces foreach -A --include example-backend --include example-app-next --parallel -v -i run start",
     "docker-build": "yarn tsc && yarn workspace example-backend build && yarn workspace example-backend build-image",
     "fix": "backstage-cli repo fix",
     "postinstall": "husky || true",
@@ -41,8 +41,8 @@
     "snyk:test:package": "yarn snyk:test --include",
     "start": "yarn workspace example-app start",
     "start-backend": "yarn workspace example-backend start",
-    "start-backend:next": "yarn workspace example-backend-next start",
-    "start:lighthouse": "yarn workspaces foreach -A --include example-backend-next --include example-app --parallel -v -i run start",
+    "start-backend:legacy": "yarn workspace example-backend-legacy start",
+    "start:lighthouse": "yarn workspaces foreach -A --include example-backend --include example-app --parallel -v -i run start",
     "start:microsite": "cd microsite/ && yarn start",
     "start:next": "yarn workspace example-app-next start",
     "storybook": "yarn --cwd storybook && yarn --cwd storybook start",
