+++ packages/app-next/README.md
@@ -4,4 +4,4 @@ This package is an example of a Backstage application using the [new frontend](.
 
 To play with it, open a terminal and run the command: `yarn start`
 
-**NOTE:** Don't forget to open a second terminal and to launch the backend or [backend-next](../../docs/backend-system/index.md) there, using `yarn start`! The frontend requires a backend to connect to.
+**NOTE:** Don't forget to open a second terminal and to launch the backend there, using `yarn start`! The frontend requires a backend to connect to.
