--- packages/backend-next/README.md
@@ -1,5 +1,0 @@
-# example-backend-next
-
-This is an example backend for [the new Backstage backend system](https://backstage.io/docs/backend-system/).
-
-Do not use this in your own projects.
