--- packages/backend-next/knip-report.md
@@ -1,12 +1,0 @@
-# Knip report
-
-## Unused dependencies (5)
-
-| Name                                             | Location     | Severity |
-| :----------------------------------------------- | :----------- | :------- |
-| @backstage/plugin-catalog-backend-module-openapi | package.json | error    |
-| @backstage/plugin-search-backend-node            | package.json | error    |
-| @backstage/plugin-permission-common              | package.json | error    |
-| @backstage/plugin-permission-node                | package.json | error    |
-| @backstage/backend-tasks                         | package.json | error    |
-
