--- packages/backend-next/package.json
@@ -1,66 +1,0 @@
-{
-  "name": "example-backend-next",
-  "version": "0.0.26-next.1",
-  "main": "dist/index.cjs.js",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "private": true,
-  "backstage": {
-    "role": "backend"
-  },
-  "homepage": "https://backstage.io",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/backstage",
-    "directory": "packages/backend-next"
-  },
-  "keywords": [
-    "backstage"
-  ],
-  "scripts": {
-    "build": "backstage-cli package build",
-    "clean": "backstage-cli package clean",
-    "lint": "backstage-cli package lint",
-    "start": "backstage-cli package start",
-    "test": "backstage-cli package test"
-  },
-  "dependencies": {
-    "@backstage/backend-defaults": "workspace:^",
-    "@backstage/backend-plugin-api": "workspace:^",
-    "@backstage/backend-tasks": "workspace:^",
-    "@backstage/catalog-model": "workspace:^",
-    "@backstage/plugin-app-backend": "workspace:^",
-    "@backstage/plugin-auth-backend": "workspace:^",
-    "@backstage/plugin-auth-backend-module-github-provider": "workspace:^",
-    "@backstage/plugin-auth-backend-module-guest-provider": "workspace:^",
-    "@backstage/plugin-auth-node": "workspace:^",
-    "@backstage/plugin-catalog-backend": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-backstage-openapi": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-openapi": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-unprocessed": "workspace:^",
-    "@backstage/plugin-devtools-backend": "workspace:^",
-    "@backstage/plugin-kubernetes-backend": "workspace:^",
-    "@backstage/plugin-notifications-backend": "workspace:^",
-    "@backstage/plugin-permission-backend": "workspace:^",
-    "@backstage/plugin-permission-backend-module-allow-all-policy": "workspace:^",
-    "@backstage/plugin-permission-common": "workspace:^",
-    "@backstage/plugin-permission-node": "workspace:^",
-    "@backstage/plugin-proxy-backend": "workspace:^",
-    "@backstage/plugin-scaffolder-backend": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-github": "workspace:^",
-    "@backstage/plugin-search-backend": "workspace:^",
-    "@backstage/plugin-search-backend-module-catalog": "workspace:^",
-    "@backstage/plugin-search-backend-module-explore": "workspace:^",
-    "@backstage/plugin-search-backend-module-techdocs": "workspace:^",
-    "@backstage/plugin-search-backend-node": "workspace:^",
-    "@backstage/plugin-signals-backend": "workspace:^",
-    "@backstage/plugin-techdocs-backend": "workspace:^"
-  },
-  "devDependencies": {
-    "@backstage/cli": "workspace:^"
-  },
-  "files": [
-    "dist"
-  ]
-}
