+++ packages/backend/knip-report.md
@@ -1,27 +1,12 @@
 # Knip report
 
-## Unused dependencies (13)
+## Unused dependencies (5)
 
-| Name                                               | Location     | Severity |
-| :------------------------------------------------- | :----------- | :------- |
-| @backstage/plugin-scaffolder-backend-module-gitlab | package.json | error    |
-| @backstage/plugin-scaffolder-backend-module-rails  | package.json | error    |
-| @backstage/plugin-azure-sites-common               | package.json | error    |
-| @backstage/plugin-tech-insights-node               | package.json | error    |
-| azure-devops-node-api                              | package.json | error    |
-| pg-connection-string                               | package.json | error    |
-| @gitbeaker/node                                    | package.json | error    |
-| better-sqlite3                                     | package.json | error    |
-| @octokit/rest                                      | package.json | error    |
-| example-app                                        | package.json | error    |
-| mysql2                                             | package.json | error    |
-| luxon                                              | package.json | error    |
-| pg                                                 | package.json | error    |
-
-## Unused devDependencies (2)
-
-| Name                             | Location     | Severity |
-| :------------------------------- | :----------- | :------- |
-| @types/express-serve-static-core | package.json | error    |
-| @types/luxon                     | package.json | error    |
+| Name                                             | Location     | Severity |
+| :----------------------------------------------- | :----------- | :------- |
+| @backstage/plugin-catalog-backend-module-openapi | package.json | error    |
+| @backstage/plugin-search-backend-node            | package.json | error    |
+| @backstage/plugin-permission-common              | package.json | error    |
+| @backstage/plugin-permission-node                | package.json | error    |
+| @backstage/backend-tasks                         | package.json | error    |
