+++ packages/backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "example-backend",
-  "version": "0.2.98-next.1",
+  "version": "0.0.26-next.1",
   "main": "dist/index.cjs.js",
   "types": "src/index.ts",
   "license": "Apache-2.0",
@@ -18,75 +18,48 @@
     "backstage"
   ],
   "scripts": {
-    "start": "backstage-cli package start",
     "build": "backstage-cli package build",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test",
     "clean": "backstage-cli package clean",
+    "lint": "backstage-cli package lint",
+    "start": "backstage-cli package start",
+    "test": "backstage-cli package test",
     "build-image": "docker build ../.. -f Dockerfile --tag example-backend"
   },
   "dependencies": {
-    "@backstage/backend-common": "workspace:^",
+    "@backstage/backend-defaults": "workspace:^",
+    "@backstage/backend-plugin-api": "workspace:^",
     "@backstage/backend-tasks": "workspace:^",
-    "@backstage/catalog-client": "workspace:^",
     "@backstage/catalog-model": "workspace:^",
-    "@backstage/config": "workspace:^",
-    "@backstage/integration": "workspace:^",
     "@backstage/plugin-app-backend": "workspace:^",
     "@backstage/plugin-auth-backend": "workspace:^",
+    "@backstage/plugin-auth-backend-module-github-provider": "workspace:^",
+    "@backstage/plugin-auth-backend-module-guest-provider": "workspace:^",
     "@backstage/plugin-auth-node": "workspace:^",
     "@backstage/plugin-catalog-backend": "workspace:^",
+    "@backstage/plugin-catalog-backend-module-backstage-openapi": "workspace:^",
+    "@backstage/plugin-catalog-backend-module-openapi": "workspace:^",
     "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "workspace:^",
     "@backstage/plugin-catalog-backend-module-unprocessed": "workspace:^",
-    "@backstage/plugin-catalog-node": "workspace:^",
     "@backstage/plugin-devtools-backend": "workspace:^",
-    "@backstage/plugin-events-backend": "workspace:^",
-    "@backstage/plugin-events-node": "workspace:^",
     "@backstage/plugin-kubernetes-backend": "workspace:^",
+    "@backstage/plugin-notifications-backend": "workspace:^",
     "@backstage/plugin-permission-backend": "workspace:^",
+    "@backstage/plugin-permission-backend-module-allow-all-policy": "workspace:^",
     "@backstage/plugin-permission-common": "workspace:^",
     "@backstage/plugin-permission-node": "workspace:^",
     "@backstage/plugin-proxy-backend": "workspace:^",
     "@backstage/plugin-scaffolder-backend": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-confluence-to-markdown": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-gitlab": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-notifications": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-rails": "workspace:^",
+    "@backstage/plugin-scaffolder-backend-module-github": "workspace:^",
     "@backstage/plugin-search-backend": "workspace:^",
     "@backstage/plugin-search-backend-module-catalog": "workspace:^",
-    "@backstage/plugin-search-backend-module-elasticsearch": "workspace:^",
     "@backstage/plugin-search-backend-module-explore": "workspace:^",
-    "@backstage/plugin-search-backend-module-pg": "workspace:^",
     "@backstage/plugin-search-backend-module-techdocs": "workspace:^",
     "@backstage/plugin-search-backend-node": "workspace:^",
     "@backstage/plugin-signals-backend": "workspace:^",
-    "@backstage/plugin-signals-node": "workspace:^",
-    "@backstage/plugin-techdocs-backend": "workspace:^",
-    "@gitbeaker/node": "^35.1.0",
-    "@octokit/rest": "^19.0.3",
-    "@opentelemetry/api": "^1.4.1",
-    "@opentelemetry/exporter-prometheus": "^0.50.0",
-    "@opentelemetry/sdk-metrics": "^1.13.0",
-    "azure-devops-node-api": "^12.0.0",
-    "better-sqlite3": "^9.0.0",
-    "dockerode": "^4.0.0",
-    "example-app": "link:../app",
-    "express": "^4.17.1",
-    "express-prom-bundle": "^7.0.0",
-    "express-promise-router": "^4.1.0",
-    "luxon": "^3.0.0",
-    "mysql2": "^3.0.0",
-    "pg": "^8.11.3",
-    "pg-connection-string": "^2.3.0",
-    "prom-client": "^15.0.0",
-    "winston": "^3.2.1"
+    "@backstage/plugin-techdocs-backend": "workspace:^"
   },
   "devDependencies": {
-    "@backstage/cli": "workspace:^",
-    "@types/dockerode": "^3.3.0",
-    "@types/express": "^4.17.6",
-    "@types/express-serve-static-core": "^4.17.5",
-    "@types/luxon": "^3.0.0"
+    "@backstage/cli": "workspace:^"
   },
   "files": [
     "dist"
